chore(deps): fastutil4k 0.1.5 -> 0.2.0
Including a fix.
